[S11P12D203-198] - 게임퀴즈 API 변경사항 반영 및 UiEvent 추가, [S11P12D203-200] - 게임 Navigation 연결

See merge request s11-webmobile1-sub2/S11P12D203!131

+++ android/project/app/src/main/java/com/chill/mallang/data/repository/remote/QuizRepository.kt
@@ -8,9 +8,9 @@ import com.chill.mallang.data.model.response.ApiResponse
 import kotlinx.coroutines.flow.Flow
 
 interface QuizRepository {
-    suspend fun getQuizIds(areaId: Int): Flow<ApiResponse<List<Int>>>
+    suspend fun getQuizIds(areaId: Long): Flow<ApiResponse<List<Long>>>
 
-    suspend fun getQuiz(quizId: Int): Flow<ApiResponse<Game01QuizData>>
+    suspend fun getQuiz(quizId: Long): Flow<ApiResponse<Game01QuizData>>
 
     suspend fun postUserAnswer(gradingUserAnswerRequest: GradingUserAnswerRequest): Flow<ApiResponse<Unit>>
